fix-修复damagePoint的类型标注错误

+++ core/envengine/common/BaseStruct/DamageCmd.py
@@ -12,4 +12,4 @@ from dataclasses_json import dataclass_json
 @dataclass
 class DamageCmd:
 	# 伤害血量
-	damagePoint: int = 0
+	damagePoint: float = 0
